refactor(proxy): extract aggregate pushdown bridge
Move aggregate transaction helpers and raw group-row reads into aggregate_pushdown.cc.
Register aggregate_pushdown.cc in the proxy build while keeping sysvar-backed mode checks in ha_lineairdb.cc.
